+++ FocusOn/package.json
@@ -32,6 +32,7 @@
     "react-image-gallery": "^1.3.0",
     "react-infinite-scroll-component": "^6.1.0",
     "react-kakao-maps-sdk": "^1.1.27",
+    "react-quill": "^2.0.0",
     "react-youtube": "^10.1.0",
     "swiper": "^11.1.14",
     "uuid": "^11.0.2",
